+++ backend/src/municipalities/municipalities.module.ts
@@ -1,10 +1,12 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { MunicipalityEntity } from './entities/municipality.entity';
+import { MunicipalitiesController } from './municipalities.controller';
 import { MunicipalitiesService } from './municipalities.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([MunicipalityEntity])],
+  controllers: [MunicipalitiesController],
   providers: [MunicipalitiesService],
   exports: [MunicipalitiesService]
 })